// src/app/lib/types.ts
export type PlayerInfo = {
  playerurl: string;
  image: string;
  firstname: string;
  lastname: string;
  teamurl: string;
  team: string;
  number: string;
  position: string;
  height: string;
  weight: string;
  lastattended: string;
  country: string;
};

export type PlayersList = PlayerInfo[];